chore(cli): promote previously experimental types
